Увеличена стабильность работы с csv источником данных.

+++ persistence-layer/src/main/java/com/ovchingus/persistence/csv/dao/ConnectionCSV.java
@@ -1,6 +1,7 @@
 package com.ovchingus.persistence.csv.dao;
 
 import com.ovchingus.persistence.GenericDao;
+import com.ovchingus.persistence.settings.DaoSettings;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -10,53 +11,78 @@ import java.io.*;
 abstract class ConnectionCSV<T> implements GenericDao<T, Integer> {
 
     private static final Logger log = LogManager.getLogger(ConnectionCSV.class);
+    String filePath = DaoSettings.getCsvFilePath();
+    String tempPath = DaoSettings.getCsvFilePath();
 
     abstract void clearFile();
 
-    boolean deleteAll(String filePath, String tempPath) {
+    boolean initialize(String filePath, String tempPath) {
         File sourceFile = new File(filePath);
         File outputFile = new File(tempPath);
-        boolean a = false;
-        boolean b = false;
+        boolean a = true;
+        boolean b = true;
         try {
-            if (!sourceFile.exists())
+            if (!sourceFile.exists()) {
                 a = sourceFile.createNewFile();
-            if (!outputFile.exists())
+                log.info("File " + filePath + " created.");
+            }
+
+            if (!outputFile.exists()) {
                 b = outputFile.createNewFile();
-            return a && b
-                    && sourceFile.delete()
-                    && outputFile.renameTo(sourceFile);
+                log.info("File " + tempPath + " created.");
+            }
+
+            if (!a || !b)
+                return false;
         } catch (IOException e) {
             e.printStackTrace();
             log.error(e.getMessage());
             return false;
-        } finally {
-            clearFile();
+        }
+
+        return true;
+    }
+
+    boolean deleteAll(String filePath, String tempPath) {
+        if (initialize(filePath, tempPath)) {
+            File sourceFile = new File(filePath);
+            File outputFile = new File(tempPath);
+            try {
+                return sourceFile.delete()
+                        && outputFile.renameTo(sourceFile);
+            } finally {
+                clearFile();
+            }
+        } else {
+            log.error("Delete all " + filePath + " rejected. Initialize error");
+            return false;
         }
     }
 
     void clear(String filePath, String tempPath) {
-        File sourceFile = new File(filePath);
-        File outputFile = new File(tempPath);
+        if (initialize(filePath, tempPath)) {
+            File sourceFile = new File(filePath);
+            File outputFile = new File(tempPath);
 
-        try (BufferedReader reader = new BufferedReader(new FileReader(sourceFile));
-             BufferedWriter writer = new BufferedWriter(new FileWriter(outputFile))) {
+            try (BufferedReader reader = new BufferedReader(new FileReader(sourceFile));
+                 BufferedWriter writer = new BufferedWriter(new FileWriter(outputFile))) {
 
-            String line;
-            while ((line = reader.readLine()) != null) {
-                if (!line.isEmpty()) {
-                    writer.write(line);
-                    writer.newLine();
+                String line;
+                while ((line = reader.readLine()) != null) {
+                    if (!line.isEmpty()) {
+                        writer.write(line);
+                        writer.newLine();
+                    }
                 }
+                reader.close();
+                writer.close();
+                if (sourceFile.delete() && outputFile.renameTo(sourceFile))
+                    log.info("File " + filePath + " cleared.");
+                else log.error("File " + filePath + " wasn`t cleared.");
+            } catch (IOException e) {
+                e.printStackTrace();
+                log.error(e.getMessage());
             }
-            reader.close();
-            writer.close();
-            if (sourceFile.delete() && outputFile.renameTo(sourceFile))
-                log.info("File " + filePath + " cleared.");
-            else log.error("File " + filePath + " wasn`t cleared.");
-        } catch (IOException e) {
-            e.printStackTrace();
-            log.error(e.getMessage());
-        }
+        } else log.error("File " + filePath + " wasn`t cleared. Initialize error");
     }
 }
